requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
python-jose[cryptography]
passlib[pbkdf2]
email-validator
python-multipart
numpy>=1.24.0 